feat: 패키지명 변경
- board 보다는 post를 다루니 post로 변경
