--- Backend/src/main/java/gwangju/ssafy/backend/domain/board/dto/DeleteCommentRequest.java
@@ -1,20 +1,0 @@
-package gwangju.ssafy.backend.domain.board.dto;
-
-import gwangju.ssafy.backend.domain.board.entity.enums.PostCategory;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-@Getter
-@Setter
-public class DeleteCommentRequest {
-
-	private Long userId;
-	private Long commentId;
-
-}
