--- Backend/src/main/java/gwangju/ssafy/backend/domain/board/repository/CommentRepository.java
@@ -1,9 +1,0 @@
-package gwangju.ssafy.backend.domain.board.repository;
-
-import gwangju.ssafy.backend.domain.board.entity.Comment;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface CommentRepository extends JpaRepository<Comment, Long> {
-}
